services/mcp-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.24.1
redis==5.0.1
pyyaml>=6.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0 